Added ToySystem (Read Summary)
The Images in the corner change according to the tag of the test toy.

In the ToyScript there is test code being used to constantly display. Remember to remove it before testing with actual toys, as the display function needs to called from the agent script.

If a null pointer exception error arises from not being able to find the images in the UI, its because they may be inactive and the script can't find inactive objects.